refactor: Move StatusConfiguration tests to its own file
